refactor: extract query validation module
Move query validation logic and validation-focused tests out of query.rs into a dedicated private module while preserving execution behavior.
