+++ src/node.rs
@@ -12,7 +12,8 @@ use crate::indexer::{latest_tx_basis_from_sdb, Indexer};
 use crate::log::{subscribe, TxLog, TxLogReader, TxLogWriter};
 use crate::memory_log::MemoryLog;
 use crate::ops::{Entid, QueryArg, TxOp};
-use crate::query::{execute_query, validate_query, QueryResult};
+use crate::query::{execute_query, QueryResult};
+use crate::query_validation::validate_query;
 use crate::schema::IdentMap;
 use crate::slate::{in_memory_slate, local_slate, remote_slate, SlateComponents};
 use edn::query::ParsedQuery;
